+++ lib/yard-sorbet/handlers/struct_handler.rb
@@ -3,7 +3,7 @@
 
 # Handles all +const+/+prop+ calls, creating accessor methods, and compiles them for later usage at the class level
 # in creating a constructor
-class YARDSorbet::StructHandler < YARD::Handlers::Ruby::Base
+class YARDSorbet::Handlers::StructHandler < YARD::Handlers::Ruby::Base
   extend T::Sig
 
   handles method_call(:const), method_call(:prop)